Show a disabled setting->description on edit
The setting->description describes what the setting does, but we are never able to see the full description in the admin panel.  We are able to see a truncated description in the index view of the settings i.e. "Email addresses separated by comma, to b[...]" but when we edit that setting, we don't see the description at all.  The description provides useful information on what the setting does and should be shown on the edit page as a disabled field, similar to the setting->name.  I made it a textarea instead of text in case the description is overly long.